Task_2/requirements.txt
notebook==7.0.6
pandas==1.5.1
seaborn==0.11.2
scikit-learn==1.1.2
matplotlib==3.9.0
numpy==1.23.5
tensorflow==2.10.0
opencv-python==4.5.5.64
torch>=1.10.0
transformers>=4.15.0
torchvision>=0.11.0
tqdm>=4.62.0
Pillow>=8.3.0